@@ -2,7 +2,7 @@ import io
 from setuptools import setup, find_packages
 
 main_ns = {}
-exec(open("dash/version.py").read(), main_ns)  # pylint: disable=exec-used
+exec(open("dash/version.py").read(), main_ns)  # pylint: disable=exec-used, consider-using-with
 
 
 def read_req_file(req_type):
@@ -23,10 +23,10 @@ setup(
         "A Python framework for building reactive web-apps. "
         "Developed by Plotly."
     ),
-    long_description=io.open("README.md", encoding="utf-8").read(),
+    long_description=io.open("README.md", encoding="utf-8").read(),  # pylint: disable=consider-using-with
     long_description_content_type="text/markdown",
     install_requires=read_req_file("install"),
-    python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*",
+    python_requires=">=3.6",
     extras_require={
         "dev": read_req_file("dev"),
         "testing": read_req_file("testing"),
@@ -36,6 +36,7 @@ setup(
             "dash-generate-components = "
             "dash.development.component_generator:cli",
             "renderer = dash.development.build_process:renderer",
+            "dash-update-components = dash.development.update_components:cli"
         ],
         "pytest11": ["dash = dash.testing.plugin"],
     },
@@ -53,12 +54,7 @@ setup(
         "Intended Audience :: Science/Research",
         "License :: OSI Approved :: MIT License",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 2",
-        "Programming Language :: Python :: 2.7",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.3",
-        "Programming Language :: Python :: 3.4",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
