@@ -1,5 +1,3 @@
-from __future__ import print_function
-
 import sys
 import os
 import uuid
@@ -11,12 +9,11 @@ import logging
 import inspect
 
 import runpy
-import future.utils as utils
 import flask
 import requests
 
 from dash.testing.errors import NoAppFoundError, TestingTimeoutError, ServerCloseError
-import dash.testing.wait as wait
+from dash.testing import wait
 
 
 logger = logging.getLogger(__name__)
@@ -42,13 +39,15 @@ def import_app(app_file, application_name="app"):
     try:
         app_module = runpy.run_module(app_file)
         app = app_module[application_name]
-    except KeyError:
+    except KeyError as app_name_missing:
         logger.exception("the app name cannot be found")
-        raise NoAppFoundError("No dash `app` instance was found in {}".format(app_file))
+        raise NoAppFoundError(
+            "No dash `app` instance was found in {}".format(app_file)
+        ) from app_name_missing
     return app
 
 
-class BaseDashRunner(object):
+class BaseDashRunner:
     """Base context manager class for running applications."""
 
     def __init__(self, keep_open, stop_timeout):
@@ -83,10 +82,10 @@ class BaseDashRunner(object):
             try:
                 logger.info("killing the app runner")
                 self.stop()
-            except TestingTimeoutError:
+            except TestingTimeoutError as cannot_stop_server:
                 raise ServerCloseError(
                     "Cannot stop server within {}s timeout".format(self.stop_timeout)
-                )
+                ) from cannot_stop_server
         logger.info("__exit__ complete")
 
     @property
@@ -110,9 +109,7 @@ class ThreadedRunner(BaseDashRunner):
     """
 
     def __init__(self, keep_open=False, stop_timeout=3):
-        super(ThreadedRunner, self).__init__(
-            keep_open=keep_open, stop_timeout=stop_timeout
-        )
+        super().__init__(keep_open=keep_open, stop_timeout=stop_timeout)
         self.stop_route = "/_stop-{}".format(uuid.uuid4().hex)
         self.thread = None
 
@@ -169,9 +166,7 @@ class ProcessRunner(BaseDashRunner):
     """
 
     def __init__(self, keep_open=False, stop_timeout=3):
-        super(ProcessRunner, self).__init__(
-            keep_open=keep_open, stop_timeout=stop_timeout
-        )
+        super().__init__(keep_open=keep_open, stop_timeout=stop_timeout)
         self.proc = None
 
     # pylint: disable=arguments-differ
@@ -202,7 +197,7 @@ class ProcessRunner(BaseDashRunner):
         logger.debug("start dash process with %s", args)
 
         try:
-            self.proc = subprocess.Popen(
+            self.proc = subprocess.Popen(  # pylint: disable=consider-using-with
                 args, stdout=subprocess.PIPE, stderr=subprocess.PIPE
             )
             # wait until server is able to answer http request
@@ -224,15 +219,11 @@ class ProcessRunner(BaseDashRunner):
                 if self.tmp_app_path and os.path.exists(self.tmp_app_path):
                     logger.debug("removing temporary app path %s", self.tmp_app_path)
                     shutil.rmtree(self.tmp_app_path)
-                if utils.PY3:
-                    # pylint:disable=no-member
-                    _except = subprocess.TimeoutExpired
-                    # pylint: disable=unexpected-keyword-arg
-                    self.proc.communicate(timeout=self.stop_timeout)
-                else:
-                    _except = Exception
-                    logger.info("ruthless kill the process to avoid zombie")
-                    self.proc.kill()
+
+                _except = subprocess.TimeoutExpired  # pylint:disable=no-member
+                self.proc.communicate(
+                    timeout=self.stop_timeout  # pylint: disable=unexpected-keyword-arg
+                )
             except _except:
                 logger.exception(
                     "subprocess terminate not success, trying to kill "
@@ -245,7 +236,7 @@ class ProcessRunner(BaseDashRunner):
 
 class RRunner(ProcessRunner):
     def __init__(self, keep_open=False, stop_timeout=3):
-        super(RRunner, self).__init__(keep_open=keep_open, stop_timeout=stop_timeout)
+        super().__init__(keep_open=keep_open, stop_timeout=stop_timeout)
         self.proc = None
 
     # pylint: disable=arguments-differ
@@ -323,7 +314,7 @@ class RRunner(ProcessRunner):
         logger.debug("start dash process with %s", args)
 
         try:
-            self.proc = subprocess.Popen(
+            self.proc = subprocess.Popen(  # pylint: disable=consider-using-with
                 args,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
@@ -342,9 +333,7 @@ class RRunner(ProcessRunner):
 
 class JuliaRunner(ProcessRunner):
     def __init__(self, keep_open=False, stop_timeout=3):
-        super(JuliaRunner, self).__init__(
-            keep_open=keep_open, stop_timeout=stop_timeout
-        )
+        super().__init__(keep_open=keep_open, stop_timeout=stop_timeout)
         self.proc = None
 
     # pylint: disable=arguments-differ
@@ -424,7 +413,7 @@ class JuliaRunner(ProcessRunner):
         logger.debug("start Dash.jl process with %s", args)
 
         try:
-            self.proc = subprocess.Popen(
+            self.proc = subprocess.Popen(  # pylint: disable=consider-using-with
                 args,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
