@@ -1,7 +1,7 @@
 from bs4 import BeautifulSoup
 
 
-class DashPageMixin(object):
+class DashPageMixin:
     def _get_dash_dom_by_attribute(self, attr):
         return BeautifulSoup(
             self.find_element(self.dash_entry_locator).get_attribute(attr), "lxml"
