@@ -3,9 +3,21 @@
 # must come before any other imports.
 __plotly_dash = True
 from .dash import Dash, no_update  # noqa: F401,E402
-from . import dependencies  # noqa: F401,E402
+from .dependencies import (  # noqa: F401,E402
+    Input,  # noqa: F401,E402
+    Output,  # noqa: F401,E402
+    State,  # noqa: F401,E402
+    ClientsideFunction,  # noqa: F401,E402
+    MATCH,  # noqa: F401,E402
+    ALL,  # noqa: F401,E402
+    ALLSMALLER,  # noqa: F401,E402
+)  # noqa: F401,E402
 from . import development  # noqa: F401,E402
 from . import exceptions  # noqa: F401,E402
 from . import resources  # noqa: F401,E402
+from . import dcc  # noqa: F401,E402
+from . import html  # noqa: F401,E402
+from . import dash_table  # noqa: F401,E402
 from .version import __version__  # noqa: F401,E402
 from ._callback_context import callback_context  # noqa: F401,E402
+from ._callback import callback, clientside_callback  # noqa: F401,E402
