@@ -18,14 +18,16 @@ const AppProvider = ({hooks}: any) => {
 AppProvider.propTypes = {
     hooks: PropTypes.shape({
         request_pre: PropTypes.func,
-        request_post: PropTypes.func
+        request_post: PropTypes.func,
+        request_refresh_jwt: PropTypes.func
     })
 };
 
 AppProvider.defaultProps = {
     hooks: {
         request_pre: null,
-        request_post: null
+        request_post: null,
+        request_refresh_jwt: null
     }
 };
 
