@@ -1,10 +1,6 @@
 import pytest
 
-import dash_html_components as html
-
-from dash import Dash
-
-from dash.dependencies import Input, Output, State
+from dash import Dash, Input, Output, State, html
 from dash.exceptions import InvalidCallbackReturnValue, IncorrectTypeException
 
 
@@ -23,17 +19,6 @@ def test_cbva001_callback_dep_types():
             html.Div(id="out3"),
         ]
     )
-
-    with pytest.raises(IncorrectTypeException) as err:
-
-        @app.callback([[Output("out1", "children")]], [Input("in1", "children")])
-        def f(i):
-            return i
-
-        pytest.fail("extra output nesting")
-
-    assert "must be `Output`, `Input`, or `State`" in err.value.args[0]
-    assert "[<Output `out1.children`>]" in err.value.args[0]
 
     with pytest.raises(IncorrectTypeException) as err:
 
@@ -62,7 +47,7 @@ def test_cbva001_callback_dep_types():
         Input("in2", "children"),
         State("state2", "children"),
     )
-    def f2(i):
+    def f3(i):
         return i
 
     # all OK with lists
@@ -71,7 +56,7 @@ def test_cbva001_callback_dep_types():
         [Input("in3", "children")],
         [State("state3", "children")],
     )
-    def f3(i):
+    def f4(i):
         return i
 
 
@@ -147,8 +132,7 @@ def test_cbva003_list_single_output(dash_duo):
 
 
 @pytest.mark.parametrize("named_out", [True, False])
-@pytest.mark.parametrize("named_in", [True, False])
-@pytest.mark.parametrize("named_state", [True, False])
+@pytest.mark.parametrize("named_in,named_state", [(True, True), (False, False)])
 def test_cbva004_named_args(named_out, named_in, named_state, dash_duo):
     app = Dash(__name__)
     app.layout = html.Div(
