@@ -1,31 +1,31 @@
 import mock
-import dash_core_components as dcc
-import dash_html_components as html  # noqa: F401
 import dash
+from dash import dcc, html  # noqa: F401
+
 from dash.development.base_component import ComponentRegistry
 
 _monkey_patched_js_dist = [
     {
         "external_url": "https://external_javascript.js",
         "relative_package_path": "external_javascript.js",
-        "namespace": "dash_core_components",
+        "namespace": "dash",
     },
     {
         "external_url": "https://external_css.css",
         "relative_package_path": "external_css.css",
-        "namespace": "dash_core_components",
+        "namespace": "dash",
     },
     {
         "relative_package_path": "fake_dcc.js",
         "dev_package_path": "fake_dcc.dev.js",
         "external_url": "https://component_library.bundle.js",
-        "namespace": "dash_core_components",
+        "namespace": "dash",
     },
     {
         "relative_package_path": "fake_dcc.min.js.map",
         "dev_package_path": "fake_dcc.dev.js.map",
         "external_url": "https://component_library.bundle.js.map",
-        "namespace": "dash_core_components",
+        "namespace": "dash",
         "dynamic": True,
     },
 ]
@@ -36,14 +36,13 @@ class StatMock(object):
 
 
 def test_external(mocker):
-    assert "dash_core_components" in ComponentRegistry.registry
-    assert "dash_html_components" in ComponentRegistry.registry
+    assert "dash" in ComponentRegistry.registry
     mocker.patch("dash.development.base_component.ComponentRegistry.registry")
-    ComponentRegistry.registry = {"dash_core_components", "dash_html_components"}
+    ComponentRegistry.registry = {"dash", "dash.dcc", "dash.html"}
 
-    mocker.patch("dash_core_components._js_dist")
-    mocker.patch("dash_core_components.__version__")
-    mocker.patch("dash_html_components._js_dist")
+    mocker.patch("dash.dcc._js_dist")
+    mocker.patch("dash.dcc.__version__")
+    mocker.patch("dash.html._js_dist")
     dcc._js_dist = _monkey_patched_js_dist  # noqa: W0212
     dcc.__version__ = "1.0.0"
 
@@ -69,14 +68,13 @@ def test_external_kwarg():
 
 
 def test_internal(mocker):
-    assert "dash_core_components" in ComponentRegistry.registry
-    assert "dash_html_components" in ComponentRegistry.registry
+    assert "dash" in ComponentRegistry.registry
     mocker.patch("dash.development.base_component.ComponentRegistry.registry")
-    ComponentRegistry.registry = {"dash_core_components", "dash_html_components"}
+    ComponentRegistry.registry = {"dash", "dash.dcc", "dash.html"}
 
-    mocker.patch("dash_core_components._js_dist")
-    mocker.patch("dash_core_components.__version__")
-    mocker.patch("dash_html_components._js_dist")
+    mocker.patch("dash.dcc._js_dist")
+    mocker.patch("dash.dcc.__version__")
+    mocker.patch("dash.html._js_dist")
     dcc._js_dist = _monkey_patched_js_dist  # noqa: W0212,
     dcc.__version__ = "1.0.0"
 
@@ -94,14 +92,33 @@ def test_internal(mocker):
             )
 
     assert resource == [
-        "/_dash-component-suites/"
-        "dash_core_components/external_javascript.v1_0_0m1.js",
-        "/_dash-component-suites/dash_core_components/external_css.v1_0_0m1.css",
-        "/_dash-component-suites/dash_core_components/fake_dcc.v1_0_0m1.js",
+        "/_dash-component-suites/" "dash/external_javascript.v1_0_0m1.js",
+        "/_dash-component-suites/dash/external_css.v1_0_0m1.css",
+        "/_dash-component-suites/dash/fake_dcc.v1_0_0m1.js",
     ]
 
     assert (
-        "fake_dcc.min.js.map" in app.registered_paths["dash_core_components"]
+        "fake_dcc.min.js.map" in app.registered_paths["dash"]
     ), "Dynamic resource not available in registered path {}".format(
-        app.registered_paths["dash_core_components"]
+        app.registered_paths["dash"]
     )
+
+
+def test_collect_and_register_resources(mocker):
+
+    app = dash.Dash(
+        __name__, assets_folder="tests/assets", assets_ignore="load_after.+.js"
+    )
+    with mock.patch("dash.dash.os.stat", return_value=StatMock()):
+        with mock.patch("dash.dash.importlib.import_module") as import_mock:
+            with mock.patch("sys.modules", {"dash_html_components": dcc}):
+                import_mock.return_value = dcc
+                app._collect_and_register_resources(
+                    [
+                        {
+                            "namespace": "dash_html_components",
+                            "relative_package_path": "dash_html_components.min.js",
+                        },
+                    ]
+                )
+                import_mock.assert_any_call("dash_html_components")
